fix(demos): use js files for the demo server code
Since ts-node does not support project references we use js files with jsdoc type annotations instead.
