+++ packages/demos/package.json
@@ -47,7 +47,6 @@
     "style-loader": "^1.0.0",
     "styled-components": "^5.2.0",
     "ts-loader": "^6.0.0",
-    "ts-node": "^8.0.0",
     "url-loader": "^2.0.0",
     "webpack": "^4.46.0",
     "webpack-cli": "^4.5.0",
