+++ packages/demos/src/feature-app-in-feature-app/integrator.node.tsx
@@ -3,7 +3,7 @@ import {createCommonJsModuleLoader} from '@feature-hub/module-loader-commonjs';
 import {FeatureAppLoader, FeatureHubContextProvider} from '@feature-hub/react';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/server';
-import {AppRendererOptions, AppRendererResult} from '../node-integrator';
+import {AppRendererOptions, AppRendererResult} from '../app-renderer';
 import {externals} from './externals';
 
 export default async function renderApp({
