+++ packages/demos/src/history-service/integrator.node.tsx
@@ -3,7 +3,7 @@ import {defineHistoryService} from '@feature-hub/history-service';
 import {defineServerRequest} from '@feature-hub/server-request';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/server';
-import {AppRendererOptions, AppRendererResult} from '../node-integrator';
+import {AppRendererOptions, AppRendererResult} from '../app-renderer';
 import {App} from './app';
 import {rootLocationTransformer} from './root-location-transformer';
