+++ packages/demos/src/server-side-rendering/integrator.node.tsx
@@ -15,7 +15,7 @@ import {
 } from '@feature-hub/serialized-state-manager';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/server';
-import {AppRendererOptions, AppRendererResult} from '../node-integrator';
+import {AppRendererOptions, AppRendererResult} from '../app-renderer';
 import {App} from './app';
 
 export default async function renderApp({
